Make it so builtin is an independently distributable package that is used.
This change also eliminates usage of the unnamed namespace. The primitive types are now defined in the builtin namespace and this is automatically imported into every quark file.

Note that the builtin namespace should probably be changed to something that includes the word quark since there are scenarios where stuff defined here should be accessible directly from host languages and the "builtin" name is not the most obvious choice. This will follow in a subsequent change because this name is hardcoded all over the compiler and this needs to be cleaned up.