Reject case-insensitive name-collisions of public identifiers
Som of our target languages (for example Ruby) put restrictions
on what identifiers are legal in different contexts. For example,
classes and modules in Ruby should start with an upper-case letter.
Other languages (that we don't support yet) may put bigger restrictions,
like total case-insensitivity. That's why we decided that it is a
name collision for two public (currently non-locals only) identifiers
to differ only in case. For example:

    // Collision
    namespace foo {}
    namespace FOO {}

    // No collision, namespace is re-opened
    namespace qux {}
    namespace qux {}

    // Collision
    class bar {}
    class BAR {}

    void test() {
        // No collision
        int bar;
        int BAR;
    }